+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 * Deprecate `chargebackProtectionLevel` and add `protectionLevel` to `Dispute` and `DisputeSearchRequest`
+* Add `networkTokenizationAttributes` to `TransactionCreditCardRequest`
+* Add `NetworkTokenizationAttributesRequest` class
+* Add validation error codes:
+  * `CREDIT_CARD_NETWORK_TOKENIZATION_ATTRIBUTE_CRYPTOGRAM_IS_REQUIRED`
 
 ## 3.18.0
 * Add `ExternalVaultCardRequest` class